WIP: Remove text-only restriction for choice widgets
It is enough that the module item has multiple choices.
This matters now that enums also are a form of multiple choice.

This does not QUITE work as written yet, because the choice widgets
are currently typed on String, and WidgetModel.getChoices() returns
a String array. Maybe the latter should be a List<T>? Need to dig
more into the API.

As things stand, there is an infinite loop if you use an enum whose
toString() method returns something other than the raw enum value.